[Archetype builder] Make the "archetype anchor" use a total order.
Make the "archetype anchor" algorithm use the same total order used to
compare dependent types in a generic signature, so that it is a
dependable total order. Tighten up the compareDependentTypes() logic
(and checking of it) a bit now that it's being used more frequently.

The weird code around GenericSignature::isCanonicalTypeInContext() is
due to a longstanding issue where

  ArchetypeBuilder::resolveArchetype(t)->getDependentType()

is not the identity function when the incoming type is a
DependentMemberType with known associated types. We want to establish
this behavior at some point going forward, in which case we can go
back to the prior implementation of
GenericSignature::isCanonicalTypeInContext().