Fix the fixit for specialized non-generic types
If a non-generic type is specialized a fixit is provided to remove the
generic arguments. The DiagnosticEngine turns a SourceRange into a
CharSourceRange by re-lexing the token at `SourceRange::End`.

The problem solved by this change occurs, if the non-generic type is
nested in an other generic type like this:

let a: GenericType<NongenericType<NongenericType>>

The lexer doesn't know that the closing angle brackets are individual
tokens leading to the removal of both of them. We can work around this
by directly specifying the start and end locations.