Mark libc math shims with always_inline attribute
We want these to always be inlined, so mark them with ... always_inline.
The compiler happened to do the right thing with them previously, but we
shouldn't depend on that.

Un-XFAILED the tests on Linux, as I think they should work correctly now.
Also added an explicit test that remainder gets inlined as well as sqrt.