[Diagnostic verify] Support conditional diagnostic verification
The '-verify' mode doesn't respect #if directive, because non-active
conditional compilation blocks might be diagnosed by the Parser, and we want
to verify that.

In this change, you can specify optional "'(' flag ')'" right after 'expect-*'.
For instance, on 'expected-error(FLAG_NAME) {{message}}', it's verified
only if '-D FLAG_NAME' is passed as the command line option.

Example usage:

  // RUN: %target-typecheck-verify-swift -D SDK_%target-sdk-name

  #if os(Linux)
  var x = linuxAPI(0) // expected-error(SDK_linux) {{}}
  #else
  var x = otherAPI(0) // expected-error(!SDK_linux) {{}}
  #endif

  // Common
  var n = commonAPI(0) // expected-error {{}}