build: avoid re-including system paths
Do not add the (default) system include path as a system include path.
Doing so results in a `-isystem /usr/.../include`.  This changes the
include ordering, resulting in the failure to include headers from the
system when building the standard library.  This breaks compilation on
exherbo which uses a slightly different directory structure
(/usr/include -> /usr/host/include -> /usr/<target>/include).  If the
paths are canonicalised during the build, we would then add
/usr/<target>/include to the CPPFLAGS, resulting in the failure to
include C headers.  Unfortunately, this is insufficient as another site
is also including the path as a SYSTEM include path.  But, this is still
strictly an improvement.