[silgen] Change one instance of switch_enum{_addr} that were using default for the some case to just use the some case.
This is the first in a series of changes to make SILGen always emit switch_enums
such that payloads are always passed as an arguments to the destination blocks.
This is important since switch_enum today in SILGen is modeled as a consuming
operation. Thus if we follow the model today where we reuse in the destination
blocks the switched upon value, we have created either a use after free or a
double consuming (depending on how the value is used).

There is no reason why we should use a default case here for these optional enums.

rdar://29791263