[ownership-verifier] Allow for trivial cases of non-trivial enums to be passed as @owned, @unowned, and @guaranteed arguments.
The logic here is that a function argument can be viewed as a PHI node in the
program graph (i.e. cfg for each function + call graph). Just like with normal
PHI nodes, we need to treat enums passed into these PHI nodes as true sum types,
but at the same time, do not want to constrain optimization when we know that
the payloaded type is trivial.

rdar://29791263