Fix "Undefined symbol" linker error with default argument for inherited initializer (#7171)
* [SILGen] Don't crash when calling a generic init with default args. (#7169)

In cases where a default value is used for a parameter with generic
type, the argument list might be empty. In that case, we don't need
to emit any arguments!

* Fix ParameterList::clone to handle deserialized defaults arguments. (#7156)

It was checking the wrong predicate, and therefore failing to mark
inherited default arguments as actually being inherited.

While here, explicitly clear out default arguments from non-inherited
cloned parameter lists. I don't think this case can come up today, but
it's better to be correct when we do hit it.

rdar://problem/30167924