[bug-reducer] Disable bug reducer when the CMAKE_GENERATOR used is not Ninja.
The bug reducer makes assumptions about the layout of the swift build directory
that may not be true in non-Ninja builds. Two examples of cmake generators that
would cause problems are VisualStudio and Xcode. I do not have time right now to
add such support, hence the disabling of the test.

rdar://30311943