[Archetype builder] Look up nested types on the requested parent.
When requesting a nested type of a potential archetype, the archetype
builder was immediately jumping to the representative of the (parent)
archetype. However, this means that we completely lost the structure
of the incoming types... one you ask for "U.AssocType" and end up with
"T.AssocType". Worse, the "representative" can vary based on the
ordering of constraints (particularly, same-type constraints), which
means that the resulting type would be somewhat unstable.

Stop doing that. When asking for a nested type, always return a type
nested within the given potential archetype, setting up the
appropriate same-type constraints with the representative's
corresponding nested type.