[Archetype builder] Suppress redundant constraints more effectively.
Fix a couple of places where we were not managing to consider
redundant constraints as 'redundant', so the canonicalized generic
signatures were not in fact minimal. Address this in several places:

* When a type is made concrete, any superclass requirements on it are
  redundant
* When a type is made concrete, any same-type relationships among its
  nested types are redundant
* Only emit potential-archetype-to-concrete constraints in the
  resulting signature for the component anchors; not for every type.

The first two bullets were existing problems that are now fixed; the
last is something I appear to have introduced with last week's
refactoring of canonicalized same-type constraints.

The recent work to make the canonicalization deterministic exacerbated
the issue considerably, leading to huge blow-up in the size of
canonical generic signatures that cost ~100k in standard library
size. This change recovers that 100k.