build: use CMake to check compiler version
CMake already detects the compiler version, and knows the flags to use
per compiler.  As the compiler itself is pure C++, it is possible to use
a non-clang compiler to build it.  As such, rely on cmake to detect the
compiler version.  This requires slightly delaying the printing of the
compiler modules to the point that the cmake compiler checks have been
performed.